src/requirements.txt
asgiref==3.5.2
Django==4.0.6
django-extensions==3.2.0
djangorestframework==3.13.1
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
